+++ modules/Students/Student.php
@@ -387,10 +387,19 @@ if ( $_REQUEST['modfunc'] === 'update'
 
 				// Create default food service account for this student.
 				// Associate with default food service account and assign other defaults.
-				DBQuery( "INSERT INTO food_service_accounts (ACCOUNT_ID,BALANCE,TRANSACTION_ID)
-					VALUES('" . $student_id . "','0.00','0');
-					INSERT INTO food_service_student_accounts (STUDENT_ID,DISCOUNT,BARCODE,ACCOUNT_ID)
+				DBQuery( "INSERT INTO food_service_student_accounts (STUDENT_ID,DISCOUNT,BARCODE,ACCOUNT_ID)
 					VALUES('" . $student_id . "','','','" . $student_id . "')" );
+
+				// Fix SQL error, Check if Account ID already exists
+				$fs_account_id_exists = DBGetOne( "SELECT 1
+					FROM food_service_accounts
+					WHERE ACCOUNT_ID='" . (int) $student_id . "'" );
+
+				if ( ! $fs_account_id_exists )
+				{
+					DBQuery( "INSERT INTO food_service_accounts (ACCOUNT_ID,BALANCE,TRANSACTION_ID)
+						VALUES('" . $student_id . "','0.00','0');" );
+				}
 
 				// Create enrollment.
 				SaveEnrollment();