+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.0.2
 -----------------
 - Fix SQL error escape Menu title in MenuReports.php & TakeMenuCounts.php
+- MySQL fix specify COUNT column name in MenuReports.php
 
 Changes in 11.0.1
 -----------------